rename utility to interactive.ts
