+++ scripts/cc-cli/src/utils/interactive.ts
